+++ server/app/src/types/database.ts
@@ -1,4 +1,4 @@
-import { User } from "./../router/api/auth/user";
+import { User } from "./../router/api/user/user.schema";
 export type DataBase = {
   Users: User;
 };
